SIRT infrequent in 1955 and 1985
The 1955-10 Official Guide to Railways has an April 1955 timetable showing only half-hourly
service to Tottenville on SIRT, while the 1957-04 Guide has an October 1956 timetable showing
service every 15 minutes as far as Great Kills. Whether the latter had begun by the end of
1955 is an open question, but circumstantial evidence from newspaper articles (regarding both
the SIRT itself and ferry schedules) suggests not.