nyctrolleys: Remove Randalls and Wards Island ferries
https://babel.hathitrust.org/cgi/pt?id=mdp.39015027871436&view=page&seq=191 says the islands were
not open to the public. https://books.google.com/books?id=ka8JAAAAIAAJ&pg=PA92 indicates as of 1920
passes were required for Randalls Island and for Wards Island outside six hours a week of visiting
time. Cudahy "Over and Back" doesn't explicitly mark them as nonpublic (as Governors/Ellis/Hart are
marked) but the text treats them with nonpublic services.

The North Brother Island ferry remains an open question, especially for the years when the island
was used for housing.