BER: on furhter study, service to schöneweide barely qualifies in 1921, so show it 1910-1920
More specificaly: Westend-Witzleben is 1.2km and in 1921 had service running
onward to Schöneweide every 20 minutes 11:00-15:00. There's a gap from 10:26 to 10:56 which
means Treptow-Schöneweide wouldn't qualify on its own. Witzleben only opened 1916 and
Westend-Charlottenburg is over 3km so would fail stop spacing, and the gap would be within
the 11:00-15:00 window by the time the trains got to the Stadtbahn, so it's possible the
Schöneweide line shouldn't be shown in 1910 & 1915 on this basis, but sources like
https://books.google.com/books?id=vfY9AQAAMAAJ&pg=RA4-PP14 suggest that there were at least
slightly more trains running before WW1 and this is such a near miss that it seems likely
that Treptow-Schöneweide qualified in its own right then.